backend/pyproject.toml
[project]
name = "voice-email-assistant"
version = "0.1.0"
description = "Voice-driven email assistant using Pipecat and Nylas"
requires-python = ">=3.10"

dependencies = [
    "pipecat-ai[anthropic,deepgram,cartesia,silero]>=0.0.60",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "nylas>=6.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
